requirements.txt
aiosendgrid==0.1.0
anyio==3.6.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
et-xmlfile==1.1.0
Flask==2.3.2
functions-framework==3.3.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.7
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3
sendgrid==6.9.0
six==1.16.0
sniffio==1.3.0
starkbank-ecdsa==1.1.1
typing_extensions==4.5.0
volia-app-models @ git+https://github.com/DimitryMelnichenkoXD/volia-app-models.git@2ee8518094b0b92862724808065b72e3ebd11c63
watchdog==3.0.0
Werkzeug==2.3.3
zipp==3.15.0
google-cloud-functions
google-cloud-storage
